requirements.txt
langchain
faiss-cpu
openai
tqdm
transformers
datasets
sentence-transformers
python-dotenv
google-generativeai
python-dotenv
scikit-learn
